plugins/DailyLuck/requirements.txt
PyMySQL
pyyaml
request